[DT][NFC] Refactor encoding utilities.
The revision shuffles the utilities to the Encoding dialect and the
Codegen dialect:

1. Move TileMxNxK struct and getEncodingInfoForMatmul method to the
  Codegen dialect (i.e., `Dialect/Codegen/*`)
2. Move isNarrowNResult to the Encoding dialect because it does not
  depend on any other dialects other than the Encoding dialect.
3. Move lowerContractionOpWithEncoding to Codegen dialect utils for the
  preparation. All the materialization logics will be moved to Codegen
  dialect; they share the utilities during the transition period.

To accomplish (3), the revision introduces ResolveEncodingInfoFn
function type, which decouple the dependency from
MaterializeEncodingTypeConvert. It is a requirement because the type
converter uses HAL while we don't want the Codegen dialect depending on
HAL. We do not need the dependency once we move all the logics to
attribute implementation.

Minor cleanups:

- Remove the `rank` argument from getEncodingInfoForMatmul. It is not
  used at all.
- Add the `static` keyword to the local `getExpandedType` function.

Note that the `lowerSetEncodingOpToPackOp` and
`lowerUnsetEncodingToUnpackOp` functions are not moved because it
requires more changes. They will be moved in a separate patch.
