+++ compiler/src/iree/compiler/Codegen/Dialect/Codegen/Utils/Utils.h
@@ -8,6 +8,7 @@
 #define IREE_COMPILER_CODEGEN_DIALECT_CODEGEN_UTILS_H_
 
 #include "iree/compiler/Codegen/Dialect/Codegen/IR/IREECodegenTypes.h"
+#include "iree/compiler/Dialect/Encoding/IR/EncodingOps.h"
 #include "llvm-c/TargetMachine.h"
 #include "llvm/ADT/SmallVector.h"
 #include "llvm/Support/raw_ostream.h"
@@ -61,6 +62,23 @@ deserializeEncodingInfo(DictionaryAttr attr);
 /// Concatenates the vectors.
 SmallVector<int64_t>
 getExpandedTileShape(const TileSwizzle::ExpandShapeType &expandShape);
+
+struct TileMxNxK {
+  int64_t M = 1;
+  int64_t N = 1;
+  int64_t K = 1;
+};
+
+MaterializeEncodingInfo
+getEncodingInfoForMatmul(Encoding::EncodingAttr encoding, TileMxNxK tileMxNxK);
+
+//===----------------------------------------------------------------------===//
+// Operation Lowering Utilities.
+//===----------------------------------------------------------------------===//
+
+FailureOr<Operation *> lowerContractionOpWithEncoding(
+    RewriterBase &rewriter, linalg::LinalgOp linalgOp, ValueRange operands,
+    bool transposeNarrowN, ResolveEncodingInfoFn getEncodingInfo);
 
 } // namespace mlir::iree_compiler::IREE::Codegen
